hopr_ticket_manager::utils

Trait UnrealizedValue

pub trait UnrealizedValue {
    // Provided method
    fn unrealized_value(
        &self,
        _channel_id: &ChannelId,
        _min_index: Option<u64>,
    ) -> Result<Option<HoprBalance>, TicketManagerError> { ... }
}

Provided Methods

fn unrealized_value( &self, _channel_id: &ChannelId, _min_index: Option<u64>, ) -> Result<Option<HoprBalance>, TicketManagerError>

Compute the total value of tickets in the channel with the given channel_id, and optionally only starting from the given first ticket index.